requirements.txt
flask
tensorflow
firebase_admin
numpy
opencv-python
pillow